add bottombar

+++ lib/data/models/restaurant.dart
@@ -71,7 +71,7 @@ class Restaurant {
       city: json["city"],
       address: json["address"],
       pictureId: json["pictureId"],
-      rating: json["rating"],
+      rating: json["rating"].toDouble(),
     );
   }
